refactor: put puzzle data into puzzle itself
