--- Assets/ScriptableObjects/PuzzleCompleteData/Puzzle 1 Complete.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 22ea451df0ace46168e45055faba2230
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
