+++ Assets/ScriptableObjects/PuzzleCompleteData/Puzzle 2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4bad3bb501afe47e6b0a92e58dd34205
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
