--- Assets/Scripts/Save/PuzzleCompleteData.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 15f9cd09202ec4f298898b1138e22974
